+++ changelog.md
@@ -3,4 +3,5 @@
 * Implement DELETE for status api
 * Fix clean up of dangling processes
 * Fix status\_url for '/ows/status/' requests
+* Add store urls for '/ows/status/' requests